// angular/src/app/shared/view-models/interfaces.ts
export interface TextLabel {
  name: string;
  label: any;
}